Updating BudgetTransactionOverview and BudgetRemainingBudgetBarChartData class, Added val updated to refresh in ApiClient, updated gradle agp

+++ shared/src/commonMain/kotlin/com/budgetbud/kmp/models/BudgetRemainingBudgetBarChartData.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 @Serializable
 class BudgetRemainingBudgetBarChartData(
     val name : String,
-    val starting_budget : Float,
-    val remaining_budget : Float,
+    val starting_budget : Double,
+    val remaining_budget : Double,
 )
