+++ shared/src/commonMain/kotlin/com/budgetbud/kmp/models/IncomeExpenseBarChartData.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class IncomeExpenseBarChartData(
     val name: String,
-    val value: Float
+    val value: Double
 )
